refactor(backend): code quality governance across app, scripts and tests
- Fix doctor endpoint path checks to align with actual repo layout and
  create runtime dirs (logs/data/workspace) in lifespan, restoring 200
- Clean 85 mypy type errors (A-class): memory, sandbox, event bus, di,
  pregel state/command, file_patch, executor interfaces, github client
- Remove dead methods and local re-imports from agent_engine.py and
  session.py (883 -> 819 lines)
- Fix real NameError in docker_sandbox inner except shadowing outer var
- Remove unused imports, fix bare excepts, B904/B006/E741/SIM102 etc
  across routes, workflow, tools, metacognition, skills registry
- Add type ignore annotations for pydantic/mypy known incompatibilities
- Fix examples/best_practices_examples.py await-outside-async syntax
- Resolve ruff findings in root benchmark/stress/health scripts
- Type-harden memory/vector stores and multi-agent integration modules

+++ app/tools/mcp_cache.py
@@ -108,7 +108,7 @@ class ToolResultCache:
             del self._cache[k]
 
     @property
-    def stats(self) -> dict[str, int]:
+    def stats(self) -> dict[str, float]:
         """Get cache statistics."""
         return {
             "size": len(self._cache),
